refactor(mt#797): split deps-rendering.ts and formatting-utilities.ts into focused sub-modules
deps-rendering.ts (638 lines) → deps-rendering-types.ts, deps-rendering-ascii.ts,
deps-rendering-graphviz.ts; barrel is now ~20 lines.

formatting-utilities.ts (636 lines) → formatting-utilities/basic-utils.ts,
formatting-utilities/config-display.ts, formatting-utilities/source-display.ts;
barrel is now ~30 lines.

All original public exports preserved.
